flash -> unity

+++ src/main/java/gforum/entities/HCommentOverview.java
@@ -8,25 +8,25 @@ import java.util.List;
 
 public class HCommentOverview implements HOverview {
 
-    private final int guildId;
+    private final long guildId;
     private final int threadId;
     private final int startIndex;
     private final List<HComment> comments;
 
 
     public HCommentOverview(HPacket packet) {
-        guildId = packet.readInteger();
+        guildId = packet.readLong();
         threadId = packet.readInteger();
         startIndex = packet.readInteger();
 
         comments = new ArrayList<>();
-        int length = packet.readInteger();
+        int length = packet.readShort();
         for (int i = 0; i < length; i++) {
             comments.add(new HComment(packet));
         }
     }
 
-    public int getGuildId() {
+    public long getGuildId() {
         return guildId;
     }
